# docs/api-contracts/openapi/client.yaml
openapi: 3.0.3
info:
  title: client-service
  version: 0.2.0
  description: 'Feature 2: Client Profile Management and Account Management.

    Agents can only access their own clients; admins can access all clients.'
servers:
- url: http://localhost
  description: Local / ingress base URL.
tags:
- name: health
- name: clients
- name: accounts
paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, example: ok}
                  service: {type: string, example: clients-service}
                required: [status]
  /api/v1/health:
    get:
      deprecated: true
      tags: [health]
      summary: (Deprecated) Health check
      responses: {'200': {description: OK}}
  /api/v1/clients/health:
    get:
      deprecated: true
      tags: [health]
      summary: (Deprecated) Clients route health check
      responses: {'200': {description: OK}}
  /api/clients:
    get:
      tags: [clients]
      summary: List clients
      description: Admins see all clients. Agents see only clients assigned to them.
      security: [{BearerAuth: []}]
      x-required-roles: [admin, agent]
      parameters:
      - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
      - {name: q, in: query, schema: {type: string, description: Optional search query (name/email/phone).}}
      responses:
        '200':
          description: List of clients.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Client'}
                  pagination: {$ref: '#/components/schemas/Pagination'}
                required: [data]
        '401':
          description: Unauthorized
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
        '403':
          description: Forbidden
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    post:
      tags: [clients]
      summary: Create client profile
      description: Creates a client profile. A log entry must be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClientCreateRequest'}
      responses:
        '201':
          description: Client created.
          content: {application/json: {schema: {$ref: '#/components/schemas/Client'}}}
        '400':
          description: Validation error.
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
        '409':
          description: Conflict (e.g., email/phone already exists).
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
        '401':
          description: Unauthorized
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
        '403':
          description: Forbidden
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  /api/clients/{clientId}:
    get:
      tags: [clients]
      summary: Get client profile
      description: Retrieves client information. A log entry should be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: clientId, in: path, required: true, schema: {$ref: '#/components/schemas/ClientId'}}
      responses:
        '200': {description: Client found., content: {application/json: {schema: {$ref: '#/components/schemas/Client'}}}}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    put:
      tags: [clients]
      summary: Update client information
      description: Updates client information. A log entry must be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: clientId, in: path, required: true, schema: {$ref: '#/components/schemas/ClientId'}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClientUpdateRequest'}
      responses:
        '200': {description: Client updated., content: {application/json: {schema: {$ref: '#/components/schemas/Client'}}}}
        '400': {description: Validation error., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: Conflict (e.g., email/phone already exists)., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    delete:
      tags: [clients]
      summary: Delete client profile
      description: Deletes a client profile. A log entry must be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: clientId, in: path, required: true, schema: {$ref: '#/components/schemas/ClientId'}}
      responses:
        '204': {description: Client deleted.}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /api/clients/{clientId}/verify:
    post:
      tags: [clients]
      summary: Verify client identity
      description: Updates client's identity verification status based on provided NRIC/document reference.
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: clientId, in: path, required: true, schema: {$ref: '#/components/schemas/ClientId'}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VerifyClientRequest'}
      responses:
        '200': {description: Verification status updated., content: {application/json: {schema: {$ref: '#/components/schemas/VerifyClientResponse'}}}}
        '400': {description: Validation error., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /api/accounts:
    post:
      tags: [accounts]
      summary: Create account
      description: Creates an account for a client. A log entry must be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AccountCreateRequest'}
      responses:
        '201': {description: Account created., content: {application/json: {schema: {$ref: '#/components/schemas/Account'}}}}
        '400': {description: Validation error., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Client not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /api/accounts/{accountId}:
    get:
      tags: [accounts]
      summary: Get account by ID
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: accountId, in: path, required: true, schema: {$ref: '#/components/schemas/AccountId'}}
      responses:
        '200': {description: Account found., content: {application/json: {schema: {$ref: '#/components/schemas/Account'}}}}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    delete:
      tags: [accounts]
      summary: Delete account
      description: Deletes an account. A log entry must be recorded (Feature 3).
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: accountId, in: path, required: true, schema: {$ref: '#/components/schemas/AccountId'}}
      responses:
        '204': {description: Account deleted.}
        '404': {description: Not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /api/clients/{clientId}/accounts:
    get:
      tags: [accounts]
      summary: List accounts for a client
      security: [{BearerAuth: []}]
      x-required-roles: [agent, admin]
      parameters:
      - {name: clientId, in: path, required: true, schema: {$ref: '#/components/schemas/ClientId'}}
      responses:
        '200': {description: Accounts for client., content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Account'}}}}}
        '404': {description: Client not found., content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT access token minted via an OAuth 2.0 provider (e.g.,
        Cognito). Each request must be authenticated and authorized (Zero Trust).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: {type: string, example: validation_error}
        message: {type: string, example: Email is required}
        requestId: {type: string, example: 7f1d2c3b-0c44-4ce9-a7ed-4c0c0c8edc1d}
      required: [error, message]
    Pagination:
      type: object
      properties:
        limit: {type: integer, minimum: 1, maximum: 200, default: 50}
        offset: {type: integer, minimum: 0, default: 0}
        total: {type: integer, minimum: 0}
    ClientId:
      type: string
      description: System-generated unique, non-editable client identifier.
      example: clt_01HZY9T3W4D7Y4W7ZB2X7MZ8QK
    AccountId:
      type: string
      description: System-generated account identifier.
      example: acc_01HZY9T3W4D7Y4W7ZB2X7MZ8QK
    Gender:
      type: string
      enum: [Male, Female, Non-binary, Prefer not to say]
    IdentityVerificationStatus:
      type: string
      enum: [unverified, pending, verified, rejected]
      default: unverified
    AccountType:
      type: string
      enum: [Savings, Checking, Business]
    AccountStatus:
      type: string
      enum: [Active, Inactive, Pending]
    Client:
      type: object
      properties:
        clientId: {$ref: '#/components/schemas/ClientId'}
        firstName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        lastName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        dateOfBirth: {type: string, format: date, description: Must be in the past; age 18-100.}
        gender: {$ref: '#/components/schemas/Gender'}
        emailAddress: {type: string, format: email}
        phoneNumber: {type: string, minLength: 10, maxLength: 15, example: '+6591234567'}
        address: {type: string, minLength: 5, maxLength: 100}
        city: {type: string, minLength: 2, maxLength: 50}
        state: {type: string, minLength: 2, maxLength: 50}
        country: {type: string, minLength: 2, maxLength: 50, example: Singapore}
        postalCode: {type: string, minLength: 4, maxLength: 10}
        identityVerificationStatus: {$ref: '#/components/schemas/IdentityVerificationStatus'}
        assignedAgentId:
          type: string
          description: Agent user id responsible for this client. Set by system from token; agents see only their own clients.
          readOnly: true
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
      required:
      - clientId
      - firstName
      - lastName
      - dateOfBirth
      - gender
      - emailAddress
      - phoneNumber
      - address
      - city
      - state
      - country
      - postalCode
      - identityVerificationStatus
    ClientCreateRequest:
      type: object
      properties:
        firstName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        lastName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        dateOfBirth: {type: string, format: date}
        gender: {$ref: '#/components/schemas/Gender'}
        emailAddress: {type: string, format: email}
        phoneNumber: {type: string, minLength: 10, maxLength: 15}
        address: {type: string, minLength: 5, maxLength: 100}
        city: {type: string, minLength: 2, maxLength: 50}
        state: {type: string, minLength: 2, maxLength: 50}
        country: {type: string, minLength: 2, maxLength: 50, example: Singapore}
        postalCode: {type: string, minLength: 4, maxLength: 10}
      required: [firstName, lastName, dateOfBirth, gender, emailAddress, phoneNumber, address, city, state, country, postalCode]
      additionalProperties: false
    ClientUpdateRequest:
      type: object
      properties:
        firstName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        lastName: {type: string, minLength: 2, maxLength: 50, pattern: '^[A-Za-z ]+$'}
        dateOfBirth: {type: string, format: date}
        gender: {$ref: '#/components/schemas/Gender'}
        emailAddress: {type: string, format: email}
        phoneNumber: {type: string, minLength: 10, maxLength: 15}
        address: {type: string, minLength: 5, maxLength: 100}
        city: {type: string, minLength: 2, maxLength: 50}
        state: {type: string, minLength: 2, maxLength: 50}
        country: {type: string, minLength: 2, maxLength: 50}
        postalCode: {type: string, minLength: 4, maxLength: 10}
      additionalProperties: false
    VerifyClientRequest:
      type: object
      properties:
        nric: {type: string, minLength: 6, maxLength: 20, description: Identity document reference (store securely).}
        documentType: {type: string, enum: [NRIC], default: NRIC}
        documentRef: {type: string, description: Reference to uploaded/secured doc (do not send raw docs via API).}
      required: [nric]
    VerifyClientResponse:
      type: object
      properties:
        clientId: {$ref: '#/components/schemas/ClientId'}
        identityVerificationStatus: {$ref: '#/components/schemas/IdentityVerificationStatus'}
      required: [clientId, identityVerificationStatus]
    Account:
      type: object
      properties:
        accountId: {$ref: '#/components/schemas/AccountId'}
        clientId: {$ref: '#/components/schemas/ClientId'}
        accountType: {$ref: '#/components/schemas/AccountType'}
        accountStatus: {$ref: '#/components/schemas/AccountStatus'}
        openingDate: {type: string, format: date}
        initialDeposit: {type: number, minimum: 0, example: 5000}
        currency: {type: string, example: SGD}
        branchId: {type: string, example: UBS-SG-001}
        createdAt: {type: string, format: date-time}
      required: [accountId, clientId, accountType, accountStatus, openingDate, initialDeposit, currency, branchId]
    AccountCreateRequest:
      type: object
      properties:
        clientId: {$ref: '#/components/schemas/ClientId'}
        accountType: {$ref: '#/components/schemas/AccountType'}
        accountStatus: {$ref: '#/components/schemas/AccountStatus'}
        openingDate: {type: string, format: date}
        initialDeposit: {type: number, minimum: 0}
        currency: {type: string, example: SGD}
        branchId: {type: string}
      required: [clientId, accountType, accountStatus, openingDate, initialDeposit, currency, branchId]
      additionalProperties: false
  responses: {}