# services/backend/clients-service/src/main/java/com/itsa/crm/clients_service/dto/VerifyClientRequest.java
package com.itsa.crm.clients_service.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public record VerifyClientRequest(
	@NotBlank
	@Size(min = 6, max = 20)
	String nric,
	@Pattern(regexp = "^NRIC$", message = "documentType must be NRIC")
	String documentType,
	String documentRef
) {}